Mockup display without real-time data
This is a part of the end product developed that showcases the basic working operations of the portal developed. Wont run as the access to real-time data is now revoked. However a json file of mockup values of coordinates and paths can be parsed to visualize some of the workings of the components.